requirements.txt
python-telegram-bot==20.7
pytz==2024.1
python-dotenv==1.0.0
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
aiohttp==3.9.3
gunicorn==21.2.0